Add git branching etiquette to README.md
Authored-by: Ben Davis bengineerdavis@gmail.com

Co-authored-by: Sean Payne sean.payne.es@gmail.com

Since new changes were made to make committing clearer and better
formatted, the README.md needed to reflect this so we could commuicate
these new standards to the rest of the team.

Changing the README both to havea  table of contents as well as a
"Git Style: How to Make Commits" section.

Slight changes were also made to "Running the application with
Docker-compose" section, to better organize the directions
for both the development and production environments.

Feedback and user testing of new instructions and formatting welcome!