Moving fork and git message ettiquette to CONTRIB
Authored-by: Ben Davis bengineerdavis@gmail.com

Co-authored-by: Sean Payne sean.payne.es@gmail.com

I'm refactoring to account for the pre-existing CONTRIB file so dev
instructions are in the right place. Earlier, we also acknowledged that
we needed an easier way to encourage verbose git messages that use
best practices, to foster better teamwork and troubleshooting.

- Adding a .gitmessagetemplate with guiding comments to facilitate easy
reference to good style and form in git messages.
- Adding a section to CONTRIB that specifies the need of local forks and
feature branching for dev.
- Adding a section that apply copies of the .gitmessagetemplate for local
feature branch commit formatting.

Please take care to use the .gitmessagetemplate for all git commits to
this repository from here on out.