Effect sources added to feed response
Effect sources added to feed response.

* added the function get_causal_sources() to score_nodes.py
* get_causal_sources is called in the function simple_scoring in
score_nodes.py
* added processing of causal effect sources to the networkx object
(added code to the function makeGraph in make_graph.py

Note that the code works but that it may look wrong when looking at the
output, but this is because the function set_edge_properties() doesn’t
behave properly. But fixing set_edge_properties() and making needed
changes to remove_edge_properties_from_nodes() functions is out of the
scope of CM-264 branch/PR