Removing files and doc descriptions for API
Co-Authored-by: Ben Davis bengineerdavis@gmail.com

Co-Authored-by: Sean Payne sean.payne.es@gmail.com

We are advertising Swagger automated docs for our API, which is
currently unsupported in our code. We're making clear in the README
that this is an issue we're coming back to shortly and avoid confusion
for those that would be chasing their tails looking for docs and
API specs from a Swagger instance that don't yet exist.

Note