Creating tests directory for app/
Co-Authored-by: Ben Davis bengineerdavis@gmail.com

Co-Authored-by: Sean Payne sean.payne.es@gmail.com

We currently do not have a location for tests within our backend
codebase. This addition will help us establish conventions for better
collaboration and codebase integrity.

Changes:
 * create app/tests
 * create subdirectory for api tests
 * create subdirectory for graph and ontology tests
 * move what ought to be a relevant test file

There are other test files I didn't move because we don't yet know if
we need them or if they in fact change the code in other unanticipated ways.