Fixed bug and logic
Bug = code was looking for value in the database that had not yet been
written to the db.

Logic missing = The top 3 shared topics are supposed to be ordered via
version 1 scoring using user b’s scores. Previously, the code was
ordering all topics with the shared personal value using version 1
scoring, which close but not what we were aiming for as outlined in
CM-921

Additional logic needed = The sorting function results changed to
returned lists instead of dictionaries because only iri are needed and
dictionaries caused errors. This is minor, but necessary because if
there are less than 3 topics possible, 1 topic needs to be repeated and
if dictionaries are used then the topic can’t be repeated because
dictionaries can’t have duplicate keys. When the ontology is expanded
this can be reverted back to dictionaries instead of lists returned.