Use test app link (#368)
* Use test app link

Use test app url for now so that beta testers aren’t directed to login to main app, but rather directed to login to the test app. This will need to be switched to production url when getting merged to master later on.

* Test site url for login (#369)

Update welcome email to direct to test website for beta testing. Will need to be changed later to direct to production website when this code gets merged to the master branch.