fix: Removes Chokidar from Docker Compose files
Attempting to run Docker Compose is throwing an error:

```
ERROR: The Compose file './docker/docker-compose.yml' is invalid because:
services.webapp.environment.CHOKIDAR_USEPOLLING contains true, which is an invalid type, it should be a string, number, or a null
```

Since it seems that Chokidar is not being using at all, this commit
removes it from the Docker Compose files.

Fixes #437.